[FIX] point_of_sale, pos_epson_printer: Correctly handle print error
When printer isn't reacheable a traceback is shown when trying to
print a receipt. This is because the error is not handled correctly.

This commit adds a check to ensure that the error is correctly
handled and that the user is notified of the error in a user-friendly
manner.

closes odoo/odoo#206457

X-original-commit: 36cfc7056ec63df4846501d301b9a0fd01e4e567
Related: odoo/enterprise#83611
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: David Monnom (moda) <moda@odoo.com>